fixed posts to update with a new post and deletes now update as well as cannot post a empty string
