+++ dvlpr/client/src/icons/Logo.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const Logo = () => {
   return (
